added seedjs file
