+++ db.js
@@ -1,5 +1,4 @@
 var mysql = require('mysql');
-var faker = require('faker');
 
 var connection = mysql.createConnection({
   host: 'localhost',
@@ -16,19 +15,6 @@ connection.connect((err) => {
 
   console.log('connected as id ' + connection.threadId);
 });
-
-const seedData = (num) => {
-  for (let i = 1; i < num; i++) {
-    connection.query(`INSERT INTO review (id, author, image, body, host_id, created_at) VALUES (${i}, '${faker.name.findName()}', '${faker.image.avatar()}', '${faker.lorem.sentence()}', 1, '2017-08-02');`, (err, data) => {
-      if (err) {
-        console.log('SEEDING: ', err);
-      } else {
-        console.log('DB successfully seeded!');
-      }
-    });
-  }
-};
-seedData(100);
 
 module.exports = {
